World map editor interactive grid

+++ DnDWorldMapEditor/Controllers/WorldMapController.cs
@@ -1,4 +1,3 @@
-
 using System.Security.Claims;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -7,6 +6,7 @@ using DnDWorldMapEditor.Models;
 using DnDWorldMapEditor.ViewModels;
 using Microsoft.IdentityModel.Tokens;
 using DnDWorldMapEditor.HelperFunctions;
+using Microsoft.AspNetCore.Authorization;
 
 namespace DnDWorldMapEditor.Controllers
 {
@@ -15,8 +15,9 @@ namespace DnDWorldMapEditor.Controllers
         private readonly ApplicationDbContext _context;
         private readonly IWebHostEnvironment _environment;
         private readonly ILogger<WorldMapController> _logger;
-        
-        public WorldMapController(ApplicationDbContext context,  IWebHostEnvironment environment, ILogger<WorldMapController> logger)
+
+        public WorldMapController(ApplicationDbContext context, IWebHostEnvironment environment,
+            ILogger<WorldMapController> logger)
         {
             _context = context;
             _environment = environment;
@@ -24,11 +25,13 @@ namespace DnDWorldMapEditor.Controllers
         }
 
         // GET: WorldMap
+        [Authorize]
         public async Task<IActionResult> Index()
         {
             try
             {
-                var worldMaps = await _context.WorldMap.ToListAsync();
+                var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+                var worldMaps = await _context.WorldMap.Where(x => x.UserId == userId).ToListAsync();
                 _logger.LogInformation("World Map Index Called, World Map Count: {worldMaps.Count}", worldMaps.Count);
                 return View(worldMaps);
             }
@@ -36,10 +39,12 @@ namespace DnDWorldMapEditor.Controllers
             {
                 _logger.LogError(ex, ex.Message);
             }
+
             return NotFound();
         }
 
         // GET: WorldMap/Details/5
+        [Authorize]
         public async Task<IActionResult> Details(int? id)
         {
             try
@@ -59,51 +64,38 @@ namespace DnDWorldMapEditor.Controllers
                     return NotFound();
                 }
 
-                var gridSpaces = _context.GridSpace.Where(x => x.WorldMapId == worldMap.Id).ToList();
-                List<GridEncounter> gridEncounters = new List<GridEncounter>();
-                List<GridCharacter> gridCharacters = new List<GridCharacter>();
-                foreach (var gridSpace in gridSpaces)
-                {
-                    gridEncounters.AddRange(_context.GridEncounter.Where(x => x.GridSpaceId == gridSpace.Id).ToList());
-                    gridCharacters.AddRange(_context.GridCharacter.Where(x => x.GridSpaceId == gridSpace.Id).ToList());
-                }
-                WorldMapDetailsViewModel viewModel = new WorldMapDetailsViewModel()
-                {
-                    WorldMap = worldMap,
-                    GridSpaces = gridSpaces,
-                    GridCharacters = gridCharacters,
-                    GridEncounters = gridEncounters,
-
-                };
-                return View(viewModel);
+                return View(worldMap);
             }
             catch (Exception ex)
             {
-                _logger.LogError("Exception Caught in WorldMap Details Get -> id:{id}\n{ex}",id ,ex.Message);
+                _logger.LogError("Exception Caught in WorldMap Details Get -> id:{id}\n{ex}", id, ex.Message);
             }
-            return NotFound();
 
+            return NotFound();
         }
 
         // GET: WorldMap/Create
+        [Authorize]
         public IActionResult Create()
         {
             return View();
         }
-        
+
         // POST: WorldMap/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Description,MapSize,BackgroundImage")] WorldMapCreateViewModel worldMapVm)
+        [Authorize]
+        public async Task<IActionResult> Create(
+            [Bind("Name,Description,MapSize,BackgroundImage")]
+            WorldMapCreateViewModel worldMapVm)
         {
             if (ModelState.IsValid)
             {
                 _logger.LogInformation("Creating WorldMap input: {0}", worldMapVm);
                 WorldMap newWorldMap = new WorldMap
                 {
-                    
                     Name = worldMapVm.Name,
                     Description = worldMapVm.Description,
                     TotalRows = 5,
@@ -121,7 +113,7 @@ namespace DnDWorldMapEditor.Controllers
                 {
                     return BadRequest("World Map must have a name.");
                 }
-                
+
                 if (User.Identity is { IsAuthenticated: true })
                 {
                     var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
@@ -155,38 +147,40 @@ namespace DnDWorldMapEditor.Controllers
                     default:
                         return BadRequest("Invalid Map Size Format. Map Size must be Small, Medium, or Large.");
                 }
-                
+
                 _context.Add(newWorldMap);
                 await _context.SaveChangesAsync();
-                
+
                 _logger.LogInformation("Created WorldMap: {0}", newWorldMap);
-                await CreateGridSpaces( newWorldMap);
+                await CreateGridSpaces(newWorldMap);
                 int gridSpaceCount = await _context.GridSpace.CountAsync(x => x.WorldMapId == newWorldMap.Id);
-                _logger.LogInformation("Created GridSpaces, Count: {0} for WorldMap: {1}", gridSpaceCount , newWorldMap.Id);
+                _logger.LogInformation("Created GridSpaces, Count: {0} for WorldMap: {1}", gridSpaceCount,
+                    newWorldMap.Id);
                 return RedirectToAction(nameof(Index));
             }
 
-            return BadRequest(ModelState); 
+            return BadRequest(ModelState);
         }
 
-        
+
         // // GET: WorldMap/Edit/5
-        [BindProperty]
-        private WorldMapEditViewModel ViewModel { get; set; }
+        [BindProperty] private WorldMapEditViewModel ViewModel { get; set; }
+
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
                 return NotFound();
             }
-        
+
             var worldMap = await _context.WorldMap.FindAsync(id);
-            
+
             if (worldMap == null)
             {
                 return NotFound();
             }
-            
+
             ViewModel = new WorldMapEditViewModel()
             {
                 Name = worldMap.Name,
@@ -194,33 +188,38 @@ namespace DnDWorldMapEditor.Controllers
                 OldImage = worldMap.BackgroundImage,
                 NewImage = null,
             };
-            
+
             return View(ViewModel);
         }
+
 
         // POST: WorldMap/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Name,Description,OldImage,NewImage")] WorldMapEditViewModel updatedMap)
+        [Authorize]
+        public async Task<IActionResult> Edit(int id,
+            [Bind("Name,Description,OldImage,NewImage")]
+            WorldMapEditViewModel updatedMap)
         {
             WorldMap? worldMap = _context.WorldMap.FirstOrDefault(m => m.Id == id);
             if (worldMap is null)
             {
                 return NotFound();
             }
-            
+
             if (ModelState.IsValid)
             {
                 try
                 {
-                    if (updatedMap.NewImage is not null )
+                    if (updatedMap.NewImage is not null)
                     {
-                        if(updatedMap.NewImage.Length != 0)
+                        if (updatedMap.NewImage.Length != 0)
                         {
                             string newFileName = FileFunctions.GenerateUniqueFileName(updatedMap.NewImage.FileName);
-                            var oldPath = Path.Combine(_environment.WebRootPath,"images", "worldMaps", worldMap.BackgroundImage);
+                            var oldPath = Path.Combine(_environment.WebRootPath, "images", "worldMaps",
+                                worldMap.BackgroundImage);
                             var newPath = Path.Combine(_environment.WebRootPath, "images", "worldMaps", newFileName);
                             worldMap.BackgroundImage = newFileName;
                             if (System.IO.File.Exists(oldPath))
@@ -233,10 +232,10 @@ namespace DnDWorldMapEditor.Controllers
                             return BadRequest("Error: New Background Image uploaded is empty.");
                         }
                     }
-                    
+
                     worldMap.Name = updatedMap.Name;
                     worldMap.Description = updatedMap.Description;
-                    
+
                     _context.Update(worldMap);
                     await _context.SaveChangesAsync();
                 }
@@ -244,12 +243,15 @@ namespace DnDWorldMapEditor.Controllers
                 {
                     _logger.LogError(ex, ex.Message);
                 }
+
                 return RedirectToAction(nameof(Index));
             }
+
             return View(worldMap);
         }
 
         // GET: WorldMap/Delete/5
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -271,16 +273,17 @@ namespace DnDWorldMapEditor.Controllers
         [HttpPost]
         [ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var worldMap = await _context.WorldMap.FindAsync(id);
             if (worldMap != null)
             {
                 var fileImage = Path.Combine(_environment.WebRootPath, "images", "worldMaps", worldMap.BackgroundImage);
-                
+
                 var gridSpaces = await _context.GridSpace.Where(x => x.WorldMapId == id).ToListAsync();
                 await DeleteGridSpaces(gridSpaces);
-                
+
                 _context.WorldMap.Remove(worldMap);
 
                 if (System.IO.File.Exists(fileImage))
@@ -290,34 +293,60 @@ namespace DnDWorldMapEditor.Controllers
             }
 
             await _context.SaveChangesAsync();
+
+
             return RedirectToAction(nameof(Index));
         }
-        
+
+        [Authorize]
         public async Task CreateGridSpaces(WorldMap worldMap)
         {
-            
             int worldMapId = worldMap.Id;
             for (int i = 0; i < worldMap.TotalRows; i++)
             {
                 for (int j = 0; j < worldMap.TotalColumns; j++)
                 {
-                    GridSpace gridSpace = new GridSpace(worldMapId,  i, j);
+                    GridSpace gridSpace = new GridSpace(worldMapId, i, j);
                     _context.GridSpace.Add(gridSpace);
                     await _context.SaveChangesAsync();
                 }
             }
-
         }
-        
+
+        [Authorize]
         public async Task DeleteGridSpaces(List<GridSpace> gridSpacesToDelete)
         {
             foreach (var gridSpace in gridSpacesToDelete)
             {
-                _context.GridSpace.Remove(gridSpace);
-                await _context.SaveChangesAsync();
-            } 
+                await DeleteGridSpace(gridSpace.Id);
+            }
         }
 
+        [Authorize]
+        public async Task DeleteGridSpace(int gridSpaceId)
+        {
+            GridSpace? gridSpace = await _context.GridSpace.FindAsync(gridSpaceId);
+            if (gridSpace != null)
+            {
+                _context.GridSpace.Remove(gridSpace);
+                await _context.SaveChangesAsync();
+                List<GridEncounter> gridEncounters =
+                    _context.GridEncounter.Where(x => x.GridSpaceId == gridSpaceId).ToList();
+                List<GridCharacter> gridCharacters =
+                    _context.GridCharacter.Where(x => x.GridSpaceId == gridSpaceId).ToList();
+
+                foreach (var encounter in gridEncounters)
+                {
+                    _context.GridEncounter.Remove(encounter);
+                    await _context.SaveChangesAsync();
+                }
+
+                foreach (var character in gridCharacters)
+                {
+                    _context.GridCharacter.Remove(character);
+                    await _context.SaveChangesAsync();
+                }
+            }
+        }
     }
-     
 }
